# lib/api-zod/src/generated/types/aiModelProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Concepful API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiModelProfile {
  id: number;
  companyId: number;
  providers?: string[];
  /** @nullable */
  modelName?: string | null;
  useCases?: string[];
  /** @nullable */
  usageNotes?: string | null;
  consentAiWorkflows?: boolean;
  consentBrandMemory?: boolean;
  updatedAt?: string;
}