# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Concepful API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  revenueRange?: string | null;
  createdAt?: string;
}