# lib/api-zod/src/generated/types/onboardingResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Concepful API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingResult {
  id: number;
  companyId: number;
  message?: string;
}